Lay out foundation for this project
This is a fairly hefty commit, but it adds all of the code which
is necessary for an MVP version of this tool, which includes one rule
whose goal is to merely to check that the `src/` directory of a project
is present.

Right now, as this package is not published, this tool may be run within
this project using the following command:

    yarn run-tool [OPTIONS] [ARGUMENTS]

There are two ways to run the tool. With no arguments, it will run the
aforementioned rule across a slew of MetaMask repositories:

    yarn run-tool

If you want to use your own list of repositories, however, you may pass
those. For instance, if you wanted to run the rule on `utils` and
`providers`, you could say:

    yarn run-tool utils providers

For help, pass `--help`:

    yarn run-tool --help

One thing about rules to note is that they are designed to be nested. So,
for instance, if you have a rule that ensures that `tsconfig.json`
contains a certain property, this rule will only work if `tsconfig.json`
exists. So you might make both rules and have the "`tsconfig.json` must
have X property" rule depend on the "`tsconfig.json` must exist" rule.
When this tool runs, it will print the results of these rules in the
same hierarchy that they were defined:

    - Does `tsconfig.json` exist? ✅
      - Does `tsconfig.json` have X property? ✅