Add file caching utility function
One of the things that this tool will do is to hit the GitHub API and
request all of the repositories under the MetaMask organization. We
don't want to do this every time the tool is run, though, or else we
might get rate limited. To prevent this, we can cache the response data
from the API in a file.

This `fetchOrPopulateFileCache` function being introduced here (which we
will use in a later commit) makes that possible.