Add function to load known MetaMask repos (#22)
If a user passes a bare identifier on the command line, like this:

```
yarn dlx @metamask/module-lint utils
```

then we assume that they want to lint the `MetaMask/utils` repository.
However we have to double-check that the repo they want to lint actually
exists. They shouldn't be able to do this, for instance:

```
yarn dlx @metamask/module-lint asdlsdfl
```

The way we do this is by pulling the list of repositories that sit under
the MetaMask GitHub organization. We exclude forks as well as archived
repos. This list is cached for an hour so that future runs of the tool
do not cause the rate limit for the GitHub API to be exceeded.