Add RepositoryFilesystem (#21)
When linting a project, we need to constantly read the files within
either the module template or the project. We want this tool to run as
fast as possible, and to do this, we can cache the metadata and content
of each file that we read so that we don't have to do it again. This
class assists with that.

---

Co-authored-by: Mark Stacey <markjstacey@gmail.com>