Add rule execution code
This commit adds code which will be used to run rules against a
particular project. This first involves looking at the dependencies
between the rules to determine the priority and order in which they
should be run, then representing the hierarchy as a tree structure.
After that, the tree is merely traversed.