Add project linting and reporting code
Once we have the ability to execute arbitrary rules on a known project,
we now need to go up one level and add the ability to receive a
reference to some kind of project (a shortname, like "utils", or a
directory path), resolve that reference to a Git repository, clone the
repository if necessary, execute the rules on that repository, and then
print out a report. This code implements those sequence of steps.