Add establishMetaMaskRepository + OutputLogger (#30)
In order to analyze a project, that project must be cloned first along
with the MetaMask module template. So, there must be some function that
checks for whether a repo has been cloned yet and clones it if not.

This commit introduces code to implement the steps involved in doing
this. It also introduces a utility class for sending messages to
standard out and standard error streams, as well as a version that can
be used in tests (which fakes out the streams).