Add remaining code to complete MVP (#33)
All of the commits up to now have been adding code to support an MVP
version of this tool. This commit ties up the MVP by adding a first
rule, which merely verifies that a project has a `src/` directory, the
entrypoint function to the tool, and the executable which kicks off the
whole thing. This commit also adds a package script, `run-tool`, which
can be used in development to quickly smoke test new changes.

---------

Co-authored-by: Mark Stacey <markjstacey@gmail.com>