Flatten rules when reporting execution results
Rules are arranged in a tree structure before being executed, so that
dependent rules are executed before their dependencies. Currently, the
report which summarizes the results of executing rules is arranged in
the same tree structure, with dependencies indented under their
dependents. While this could be interesting to some users, it is not
necessary. This commit flattens the list when reporting.