Add checks for Node version compliance (#41)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
For any given project, we want to verify the following to make sure node
version compliance:

* .nvmrc should exist, and its contents should match the same file in
the module template
* The engines.node property in package.json should exist, and should
match the same value as in the module template
* The README should include a line - Install the current LTS version of
[Node.js](https://nodejs.org).

Fixes #12