lint related dependencies conform (#45)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
We want to make sure that for a given project:

* All of the lint-related packages listed in the module template's
dependencies are present
* For each package, the version used in the project matches that in the
module template

Fixes #9