Typescript Compliance (#55)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
We want to make sure that for a given project:

- The project contains a tsconfig.json file, and the content of the file
matches the same file in the module template
- The project contains a tsconfig.build.json file, and the content of
the file matches the same file in the module template
- The project contains a tsup.config.ts file, and the content of the
file matches the same file in the module template
- The project contains @types/node, ts-node, tsup, and typescript as dev
dependencies, and the versions match the same dev dependencies as in the
module template
- The project contains the same build and build:types package scripts as
in the module template
- The project's package.json has an exports field, and all of the keys
in exports should be present in the project's exports field and the
values should match.
- The project's package.json has a main field, and its value should
match the same field in the module template.
- The project's package.json has a module field, and its value should
match the same field in the module template.
- The project's package.json has a types field, and its value should
match the same field in the module template.
- The project's package.json has a files field, and its value should
match the same field in the module template.
If project's package.json contains a lavamoat field, and that contains
allowScripts, then a tsup>esbuild property should be present and should
match the same one in the module template.

Fixes #8