typedoc compliance (#58)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
We want to make sure that for a given project:
* The project contains typedoc as a dev dependency, and the versions
match the same dev dependencies as in the module template.
* The project contains a typedoc.json, and the content of this file
matches the same as in the module template.
* The project contains the same build:docs package script as in the
module template.

Fixes [#47](https://github.com/MetaMask/module-lint/issues/47)