trimming the content removed in readFile (#62)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
I was working on the changelog.md file validation and found that
`validateChangelog` was failing with one character less! When I traced
the reason, I found that there's a trim on the content of the file in
`readFile` function. Once I removed trim, it worked fine.

I think, when we override the functionality, in this case `readFile`
overrides functionality from `metamask/utils` which uses `fs`. We need
to add an additional parameters which can modify the original behaviour,
so that consumer is aware and can utilise according to the need.