Added changelog related checks. (#63)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
We want to make sure that for a given project:

- The project contains a CHANGELOG.md, and it is valid (i.e., running it
through auto-changelog's validate function passes).
- We would also want to make sure that module-lint is using the same
version of auto-changelog as the module template.
- The project contains @metamask/auto-changelog as a dev dependency, and
the version matches the same dev dependency as in the module template.
- The project contains the same lint:changelog package script as in the
module template
- If the lint script exists, then the lint script runs yarn
lint:changelog (i.e. the value of this script matches /\byarn
lint:changelog\b/).

* Fixes #48