checks for .editorconfig .gitattributes .gitignore (#67)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
We want to make sure that for a given project:

.editorconfig is present and matches the same content as in the module
template
.gitattributes is present and matches the same content as in the module
template
.gitignore is present, and the project's version of the file starts with
the template's version

* Fixes #61