non zero exit in case of lint fails or errors (#76)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
we need the tool to exit with a non-zero exit code if there are any lint
errors or if the tool throws an error for any reason. This way, when
running the tool in a GitHub workflow, the job will fail.

* Fixes #75