generate module lint report by schedule and post to slack when failed (#81)
<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?

Are there any issues or other links reviewers should consult to
understand this pull request better? For instance:

* Fixes #12345
* See: #67890
-->
A module-lint report to be automatically generated once in a week and
shared on a slack channel by tagging the team (when failed).

For the purpose of testing, this repo has been
[forked](https://github.com/MetaMask/module-lint-fork) and all the
changes of this PR are available in the forked repo as well. And right
now, it posts to the channel `temp-test-module-lint`. Please feel free
to ping me, if you want to be added to this channel to see the scheduled
messages posted from github actions to the slack channel.

Here is the screenshot of the report posted to the slack channel.
<img width="634" alt="Screenshot 2024-04-16 at 17 29 39"
src="https://github.com/MetaMask/module-lint/assets/45172858/f0f789ac-01e8-4a68-87a5-286b635938a8">

And [click here](https://github.com/MetaMask/module-lint-fork/actions)
to see the github action (cron job) in the forked repo.

Fixes #60