Don't abort when rule throws, but continue
If a project is being linted and a rule throws an error — perhaps it
relies on a file to be in a certain format, or some other reason — then
linting for that project will halt, and the lint report for that project
will only show that error. An error like this signifies a bug in
`module-lint`, but sometimes we can't fix that bug right away, and in
the meantime this makes the report useless, since we don't know whether
any of the other rules passed or not.

This commit fixes this problem by catching errors inside of rules and
displaying them as such. To implement this change, we needed to convert
the binary status of a rule to a trinary status ("passed", "failed",
"errored"). We also updated the output of the lint report to display
this error status and distinguish it from a failed status.