Use Slack threads to produce periodic report
Currently, the GitHub workflow that runs `module-lint` once a week
posts each package's lint report in a separate Slack message. This works
fine for now because there are only a few packages we are running
`module-lint` against, but once we scale this up, it will create a lot
of noise.

To combat this, this commit updates the workflow to produce a summary
message and then post all of the per-package messages in a thread.