Update report output format (#90)
This commit simply aligns the output so that all of the icons that mark
whether a rule passed, failed, or errored are on the left hand side.
This makes the output prettier and makes it easier to tell what the
status of a rule is. This commit also eschews emoji (which pop out too
much) for other kinds of characters.