Add documentation (#89)
When this tool is run on a cadence and reports show up in Slack, people
may be confused how to interpret failures or may want to learn more
about what the output means in general. This document provides a brief
explanation of the tool, how it works, and how to take action on
reports.

Fixes #85.